LockRepoCommandHandler
