+++ src/OctoshiftCLI.Tests/ado2gh/Commands/LockRepoCommandTests.cs
@@ -13,7 +13,7 @@ namespace OctoshiftCLI.Tests.AdoToGithub.Commands
         private readonly Mock<AdoApiFactory> _mockAdoApiFactory = TestHelpers.CreateMock<AdoApiFactory>();
         private readonly Mock<OctoLogger> _mockOctoLogger = TestHelpers.CreateMock<OctoLogger>();
 
-        private readonly LockRepoCommand _command;
+        private readonly LockRepoCommandHandler _command;
 
         private const string ADO_ORG = "FooOrg";
         private const string ADO_TEAM_PROJECT = "BlahTeamProject";
@@ -24,21 +24,22 @@ namespace OctoshiftCLI.Tests.AdoToGithub.Commands
 
         public LockRepoCommandTests()
         {
-            _command = new LockRepoCommand(_mockOctoLogger.Object, _mockAdoApiFactory.Object);
+            _command = new LockRepoCommandHandler(_mockOctoLogger.Object, _mockAdoApiFactory.Object);
         }
 
         [Fact]
         public void Should_Have_Options()
         {
-            Assert.NotNull(_command);
-            Assert.Equal("lock-ado-repo", _command.Name);
-            Assert.Equal(5, _command.Options.Count);
+            var command = new LockRepoCommand(_mockOctoLogger.Object, _mockAdoApiFactory.Object);
+            Assert.NotNull(command);
+            Assert.Equal("lock-ado-repo", command.Name);
+            Assert.Equal(5, command.Options.Count);
 
-            TestHelpers.VerifyCommandOption(_command.Options, "ado-org", true);
-            TestHelpers.VerifyCommandOption(_command.Options, "ado-team-project", true);
-            TestHelpers.VerifyCommandOption(_command.Options, "ado-repo", true);
-            TestHelpers.VerifyCommandOption(_command.Options, "ado-pat", false);
-            TestHelpers.VerifyCommandOption(_command.Options, "verbose", false);
+            TestHelpers.VerifyCommandOption(command.Options, "ado-org", true);
+            TestHelpers.VerifyCommandOption(command.Options, "ado-team-project", true);
+            TestHelpers.VerifyCommandOption(command.Options, "ado-repo", true);
+            TestHelpers.VerifyCommandOption(command.Options, "ado-pat", false);
+            TestHelpers.VerifyCommandOption(command.Options, "verbose", false);
         }
 
         [Fact]
